requirements.txt
# slack
requests
python-daemon
pyyaml
websocket-client
slackclient
Pyro4

# gensim
cython
numpy
-e git+https://github.com/scipy/scipy#egg=scipy-dev
-e git+https://github.com/kelp404/six#egg=six-dev
gensim

# keras
keras
simples3

# nltk... yayyy
nltk
ipython
